+++ Examples/Sources/AppDelegate.swift
@@ -10,17 +10,10 @@ import UIKit
 
 @main
 class AppDelegate: UIResponder, UIApplicationDelegate {
-
-    var window: UIWindow?
     
     func application(_: UIApplication, didFinishLaunchingWithOptions _: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
         
         SettingsBundleHelper.applySettings(customKeysAndValues: customKeysAndValues())
-
-        window = UIWindow(frame: UIScreen.main.bounds)
-        
-        window?.rootViewController = UIStoryboard(name: "Main", bundle: nil).instantiateInitialViewController()
-        window?.makeKeyAndVisible()
         
         return true
     }